# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PeraQuest MVP API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "Healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "const": "ok"
                    }
                  },
                  "required": [
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/students/onboarding": {
      "post": {
        "operationId": "onboardStudent",
        "summary": "Create an Eiken Grade 3 learner profile and determine guardian requirements.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentOnboardingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Student onboarded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentOnboardingResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/v1/me/guardian-link": {
      "get": {
        "operationId": "getGuardianLink",
        "summary": "Read the current learner's guardian-link state.",
        "parameters": [
          {
            "$ref": "#/components/parameters/StudentId"
          }
        ],
        "responses": {
          "200": {
            "description": "Guardian-link state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuardianLink"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/me/consents/voice-processing": {
      "put": {
        "operationId": "putVoiceConsent",
        "summary": "Record or withdraw the current voice-processing consent version.",
        "parameters": [
          {
            "$ref": "#/components/parameters/StudentId"
          },
          {
            "$ref": "#/components/parameters/GuardianId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceConsentWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Consent recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceConsent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/me/capabilities": {
      "get": {
        "operationId": "getCapabilities",
        "summary": "Resolve server-authoritative learning, voice, commerce, notification, LINE-return, and entitlement capabilities.",
        "parameters": [
          {
            "$ref": "#/components/parameters/StudentId"
          },
          {
            "$ref": "#/components/parameters/ClientPlatform"
          }
        ],
        "responses": {
          "200": {
            "description": "Resolved capabilities",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Capabilities"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/me/voice-upload-ticket": {
      "post": {
        "operationId": "createVoiceUploadTicket",
        "summary": "Request a signed upload ticket after the server repeats guardian and Consent Gate checks.",
        "parameters": [
          {
            "$ref": "#/components/parameters/StudentId"
          }
        ],
        "responses": {
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "501": {
            "description": "Gate passed, but object storage is not configured in this slice",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/trial-sessions": {
      "post": {
        "operationId": "createTrialSession",
        "responses": {
          "200": {
            "description": "Trial session created"
          },
          "409": {
            "description": "Trial already redeemed"
          }
        }
      }
    },
    "/v1/me/trial-sessions/{sessionId}/answers": {
      "post": {
        "operationId": "submitTrialAnswer",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Answer feedback"
          }
        }
      }
    },
    "/v1/me/trial-sessions/{sessionId}/complete": {
      "post": {
        "operationId": "completeTrialSession",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trial completed without durable progress"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "StudentId": {
        "name": "X-Student-Id",
        "in": "header",
        "required": true,
        "description": "Scaffold-only identity header; replace with verified bearer-token middleware.",
        "schema": {
          "type": "string"
        }
      },
      "GuardianId": {
        "name": "X-Guardian-Id",
        "in": "header",
        "required": false,
        "description": "Scaffold-only linked guardian identity, required when a minor's consent is written. Replace with verified actor claims.",
        "schema": {
          "type": "string"
        }
      },
      "ClientPlatform": {
        "name": "X-Client-Platform",
        "in": "header",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/ClientPlatform"
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Guardian or voice consent gate denied the operation",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Student not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "ClientPlatform": {
        "type": "string",
        "enum": [
          "ios",
          "android",
          "pc"
        ]
      },
      "GuardianLinkStatus": {
        "type": "string",
        "enum": [
          "not_required",
          "pending",
          "verified",
          "rejected",
          "revoked"
        ]
      },
      "ConsentStatus": {
        "type": "string",
        "enum": [
          "missing",
          "granted",
          "denied",
          "withdrawn",
          "outdated"
        ]
      },
      "StudentOnboardingRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "birthMonth": {
            "type": "string",
            "pattern": "^\\d{4}-(0[1-9]|1[0-2])$"
          },
          "targetExam": {
            "const": "eiken_grade_3"
          },
          "authProvider": {
            "type": "string",
            "enum": [
              "apple",
              "google",
              "email_magic_link"
            ]
          },
          "client": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "platform": {
                "$ref": "#/components/schemas/ClientPlatform"
              },
              "deviceId": {
                "type": "string"
              },
              "appVersion": {
                "type": "string"
              },
              "osVersion": {
                "type": "string"
              }
            },
            "required": [
              "platform"
            ]
          }
        },
        "required": [
          "birthMonth",
          "targetExam",
          "authProvider",
          "client"
        ]
      },
      "StudentOnboardingResponse": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "format": "uuid"
          },
          "isMinor": {
            "type": "boolean"
          },
          "guardianLinkStatus": {
            "$ref": "#/components/schemas/GuardianLinkStatus"
          },
          "onboardingStatus": {
            "type": "string",
            "enum": [
              "pending_guardian",
              "active"
            ]
          }
        },
        "required": [
          "studentId",
          "isMinor",
          "guardianLinkStatus",
          "onboardingStatus"
        ]
      },
      "GuardianLink": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/GuardianLinkStatus"
          },
          "purchaseAllowed": {
            "type": "boolean"
          },
          "verifiedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        },
        "required": [
          "status",
          "purchaseAllowed",
          "verifiedAt"
        ]
      },
      "VoiceConsentWrite": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "granted",
              "denied",
              "withdrawn"
            ]
          },
          "version": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "status",
          "version"
        ]
      },
      "VoiceConsent": {
        "type": "object",
        "properties": {
          "type": {
            "const": "voice_processing"
          },
          "status": {
            "$ref": "#/components/schemas/ConsentStatus"
          },
          "version": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "type",
          "status",
          "version"
        ]
      },
      "Capabilities": {
        "type": "object",
        "properties": {
          "examLevel": {
            "const": "eiken_grade_3"
          },
          "platform": {
            "$ref": "#/components/schemas/ClientPlatform"
          },
          "canLearn": {
            "type": "boolean"
          },
          "canUploadVoice": {
            "type": "boolean"
          },
          "voiceUploadMode": {
            "type": "string",
            "enum": [
              "disabled",
              "signed_upload"
            ]
          },
          "canPurchase": {
            "type": "boolean"
          },
          "guardianLinkStatus": {
            "$ref": "#/components/schemas/GuardianLinkStatus"
          },
          "voiceConsentStatus": {
            "$ref": "#/components/schemas/ConsentStatus"
          },
          "consentVersionRequired": {
            "type": "string"
          },
          "paymentChannels": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "apple_app_store",
                "google_play",
                "web_checkout"
              ]
            }
          },
          "notificationChannels": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "ios_push",
                "android_push",
                "web_push",
                "line"
              ]
            }
          },
          "lineReturnTargets": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "app_deep_link",
                "web_https"
              ]
            }
          },
          "entitlements": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "examLevel",
          "platform",
          "canLearn",
          "canUploadVoice",
          "voiceUploadMode",
          "canPurchase",
          "guardianLinkStatus",
          "voiceConsentStatus",
          "consentVersionRequired",
          "paymentChannels",
          "notificationChannels",
          "lineReturnTargets",
          "entitlements"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "additionalProperties": true
      }
    }
  }
}